redistribute IMS int tests and add resource agent client tests
